backend/app/main.py
"""
FastAPI Main Application
AI-Driven Financial Assistant Backend
"""
from fastapi import FastAPI, Depends, HTTPException, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from contextlib import asynccontextmanager
import os
from dotenv import load_dotenv

from app.database import engine, Base
from app.routers import goals, budgeting, recommendations, alerts, auth, chat, obligations
from app.core.config import settings

load_dotenv()

# Create database tables
@asynccontextmanager
async def lifespan(app: FastAPI):
    # Startup
    Base.metadata.create_all(bind=engine)
    yield
    # Shutdown
    pass

app = FastAPI(
    title="AI Financial Assistant API",
    description="AI-driven financial assistant with goal planning, budgeting, and recommendations",
    version="1.0.0",
    lifespan=lifespan
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "http://localhost:8501"],  # Next.js and Streamlit
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Include routers
app.include_router(auth.router, prefix="/api/v1/auth", tags=["Authentication"])
app.include_router(goals.router, prefix="/api/v1/goals", tags=["Goals"])
app.include_router(budgeting.router, prefix="/api/v1/budgeting", tags=["Budgeting"])
app.include_router(recommendations.router, prefix="/api/v1/recommendations", tags=["Recommendations"])
app.include_router(alerts.router, prefix="/api/v1/alerts", tags=["Alerts"])
app.include_router(chat.router, prefix="/api/v1/chat", tags=["Chat"])
app.include_router(obligations.router, prefix="/api/v1/obligations", tags=["Obligations"])

@app.get("/")
async def root():
    return {
        "message": "AI Financial Assistant API",
        "version": "1.0.0",
        "docs": "/docs"
    }

@app.get("/health")
async def health_check():
    return {"status": "healthy"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)

