backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.11
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
llama-index==0.9.20
pinecone-client==2.2.4
faiss-cpu==1.7.4
numpy==1.23.5
pandas==1.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
alembic==1.12.1
